Use network mode `host` in integration docker container
The reason (I think it is important to understand) for this, is that JDBC sub connections are
going to return dynamically exposed ports. However, if other network modes used, we do not know
which ports are mapped at host for those dynamically exposed ports within a container. Therefore,
we use `--net=host` so that exposed ports are reachable and same on host.