Fix order bug in schema pruning
When selecting only required columns from schema, the new schema fields should be ordered as
required columns.

Previously, it was not ordered because of filtering a Set of fields.

This was the reason `show` or `take` actions were failing. The JDBC result set order was not
aligned with newSchema (pruned for only required columns in wrong order).