Use handle in jdbc subconnections
Instead of running execute query on each sub-connection separately, we run it in main connection
upfront. Then, each sub connection only requests the result set using a handle.

One important knowledge is that if the handle is -1 then this means empty result set and we return
empty iterator.

This should fix #22