Spark count action pushdown
If there is a `.count` action after loading the Exasol dataframe, then we should not shuffle all
those data through network just for counting afterwards.

We detect `df.count()` operation when the list of column names (`requiredColumns`) is empty. In
such cases, perform a single count star query on Exasol and create Spark RDD of empty rows as many
as resulted count.

Fixes #24