Update `repartitionPerNode` function
The problem:

If we have more Exasol data nodes than the number of Dataframe partitions, we will open more
sub-connections and some of them will not be used at all. This currently is a problem, since all
opened sub-connections should be connected and closed.

The solution:

It is easy to perform repartition on dataframe and increase the number of partitions. This is what
current commit does.

However, `repartition` is expensive operation that might involve shuffle. Therefore, we need to
check if we can use fewer (as number of partitions) when initiating sub-connections.